Move Sanitize() to internal
